value(myfig) produces an object not a number due to matlab update

+++ Protocols/@AthenaDelayComp/StimulusSection.m
@@ -122,7 +122,7 @@ switch action,
     y=5;
     SoloParamHandle(obj, 'existing_numClassPsych', 'value', 0, 'saveable', 0);
     SoloParamHandle(obj, 'existing_numClass', 'value', 0, 'saveable', 0);
-    SoloParamHandle(obj, 'my_window_info', 'value', [x, y, value(myfig)], 'saveable', 0);
+    SoloParamHandle(obj, 'my_window_info', 'value', [x, y, double(value(myfig))], 'saveable', 0);
     NumeditParam(obj,'numClass',4,x,y,'label','numClass','TooltipString','Number of stimulus pairs');
     
     set_callback_on_load(numClass, 4); %#ok<NODEF>
